+++ src/ECS/Components/attackRateComponent.rs
@@ -1,10 +1,10 @@
 
 pub struct AttackRateComponent{
-    attackRate: usize,
+    attackRate: f32,
 }
 
 impl AttackRateComponent{
-    pub fn new(atkRate: usize) -> Self {
+    pub fn new(atkRate: f32) -> Self {
         Self{
             attackRate: atkRate,
         }
@@ -14,19 +14,19 @@ impl AttackRateComponent{
 
     }
     
-    pub fn getAttackRate(&self) -> usize{
+    pub fn getAttackRate(&self) -> f32{
         return self.attackRate;
     }
 
-    pub fn setAttackRate(&mut self, attackRate: usize){
+    pub fn setAttackRate(&mut self, attackRate: f32){
         self.attackRate = attackRate;
     }
 }
 
 
 pub fn test(){
-    let mut attackComponent = AttackRateComponent{attackRate:1};
+    let mut attackComponent = AttackRateComponent{attackRate:1.0};
     println!("AttackRate at creation: {}",attackComponent.getAttackRate());
-    attackComponent.setAttackRate(2);
+    attackComponent.setAttackRate(2.0);
     println!("AttackRate after change: {}",attackComponent.getAttackRate());
 }
